# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ollama Load Balancer API",
    "description": "REST API для интеллектуального балансировщика нагрузки Ollama с мониторингом ресурсов GPU/CPU/RAM/Disk и умным распределением запросов.\n\n## Особенности\n- Умная балансировка на основе доступных ресурсов\n- Model Affinity - направление запросов к серверам с загруженной моделью\n- Real-time мониторинг через WebSocket\n- Health Check - автоматическое обнаружение нерабочих бэкендов\n- Session Stickiness - сохранение сессии на одном сервере\n- Queue Manager - обработка перегрузок с очередью запросов\n\n## Аутентификация\nAPI использует токены для аутентификации. Токен передается в заголовке `X-API-Token`.\nMaster token (первый в списке) может создавать и отзывать другие токены.",
    "version": "1.0.0",
    "contact": {
      "name": "Ollama Load Balancer"
    },
    "license": {
      "name": "MIT License"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Local development server"
    },
    {
      "url": "https://localhost:8443",
      "description": "Production server (HTTPS)"
    },
    {
      "url": "http://{host}:{port}",
      "description": "Custom server",
      "variables": {
        "host": {
          "default": "localhost",
          "description": "Hostname or IP address"
        },
        "port": {
          "default": "8081",
          "description": "API port"
        }
      }
    }
  ],
...
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/auth/token": {
      "post": {
        "tags": ["Auth"],
        "summary": "Create token",
        "description": "Создание нового API токена. Требует master token.\n\nMaster token - первый токен в списке конфигурации.",
        "operationId": "createToken",
        "responses": {
          "201": {
            "description": "Token created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - master token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Auth"],
        "summary": "Revoke token",
        "description": "Отзыв (удаление) API токена. Требует master token.\n\nMaster token нельзя отозвать.",
        "operationId": "revokeToken",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRevokeRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "description": "Токен для отзыва (альтернатива body)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token revoked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Token revoked successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - master token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Token",
        "description": "API токен для аутентификации"
      }
    },
    "schemas": {
...
      "TokenResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "token": {
            "type": "string",
            "example": "a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0u1v2w3x4y5z6"
          },
          "message": {
            "type": "string",
            "example": "Token generated successfully"
          }
        }
      },
      "TokenRevokeRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Токен для отзыва"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Unauthorized: valid API token required"
          },
          "code": {
            "type": "string",
            "description": "Код ошибки (опционально)"
          }
        }
      }
    }
  }
}